backend/requirements_new.txt
# Core FastAPI
fastapi==0.110.1
uvicorn==0.25.0
starlette==0.37.2
python-dotenv==1.1.1
pydantic==2.11.7
pydantic_core==2.33.2

# Database
motor==3.3.1
pymongo==4.5.0

# File handling
python-multipart==0.0.20

# Audio Processing (essential only)
librosa==0.11.0
soundfile==0.13.1
numpy==1.26.4
scipy==1.16.2

# AI/ML for audio processing
basic-pitch==0.4.0

# Music notation
music21==9.7.1
pretty_midi==0.2.10
mido==1.3.3

# AI Integration
emergentintegrations==0.1.0

# Utilities
python-dateutil==2.9.0.post0
